+++ CHANGELOG.md
@@ -12,3 +12,5 @@ Second version after feedback
 - Update: grayout hashtag in request id
 - Feature: set date and time of service request creation
 - Update: add shadcn calendar inputs to every date time input, and enable manual date input
+- Update: change the order of blocks in service request
+-